Initial commit
This commit just adds the README.md file which contains all the explanations
I would have written here.

They are in the file instead, because normal people see that first, then the
git history.

So if you've got so far, jump to the next lesson with the command provided in
the README.

Good luck!