The build system first
In the first commit, we just set up the basic build and test infrastructure.

We first create two subdirectories, `src/` and `tests/`. `src/` is going to
contain production code, `tests/` is going to contain the tests. We do this
because it is very important to separate test code from production code. The
worst thing you could have is test logic inside the production code - for
instance by testing if a specific variable has a magic value `"testing"`. Now,
some people prefer to have at least the unit tests closer to the corresponding
unit, but this project is for learning and the best way to see the clear
separation between test and production code is, in my opinion, by having
completely different directories right at the top.

The tests are divided in two, `tests/src/` and `tests/integration/`. For each
test suite we are going to have one executable, but since this project is so
simplistic in nature, we are going to have only two executables which will
contain the tests: one for unit tests, and one for integration tests.

As for the tests themselves, we have just created two empty tests which both
test that `1 = 1`. These are not proper unit and integration tests
respectively, but we do not have any production code to test and we don't want
to have any production code just yet; all we want is to test the build system,
the test runner, and whether the executables are linked together properly or
not.

For production code, we create a `src/main.cpp` file. We leave it empty,
because in the tests, we are not testing any production code yet. The cleanest
way would have been to not even create a `main.cpp` file, but then we would
have not been able to link against the tictactoe library, and we wanted to link
against it in order to test the build system itself.

As you can see, an empty tictactoe library suffices for this purpose. In the
following commits we are going to remove `main.cpp`, since we're going to write
some production code.

One general idea in TDD is to only write the most minimal code in order to make
the tests pass. In this particular commit, we are running three tests: two
automatic tests (one "unit" test, one "integration" test) and one manual test:
running the build and test system itself.

The infrastructure presented here uses CMake to build the corresponding
workspace for your environments. It can generate workspaces or projects based
on Unix makefiles and various editors and IDEs like Kate or Visual Studio on
Microsoft operating systems. Install CMake and run `cmake --help` from the
command line for the complete list available on your environment.

I am also using the boost test library for organizing tests. I think it is not
the best choice for learning TDD, because it does too much, but it makes the
test code look just a little bit cleaner, and it also works with minimal setup
on my system.

To generate the make files, type in the root directory of the project `cmake
.`. After that, build the default target - on the command line of a Linux
system this is accomplished with the command `make`. Then you can run the test
target (`make test` on the command line).

Now go on reading the next lesson with the command provided in the README file.