Begin the TDD cycle by writing an integration test first
First let me begin with a fact: there are different schools of thought
concerning TDD and the terminology used, and every single developer has
a different view on this matter. What is important to you, while learning, is
to use a consistent system.

I am going to use a simplified version of TDD, just enough to get across as
many ideas as possible on top of such a simple project like tic tac toe.

One way of organizing tests is in layers of abstraction or in the width of the
system under testing (SUT). The most elementar type of test is a unit test,
which tests just a unit. A unit is a public method of a class.

Above this layer are the integration tests which test the interaction of
multiple classes. There are multiple types of tests which are integrative in
nature, like component tests, subcutaneous tests, system testing, end to end
testing, and so on.

This simple project does not have the required complexity to make it worth
introducing this kind of granularity in test organization, which is why the
tests are split only in two layers, unit and integration.

Conceptually, integration tests depend on unit tests, because we should not
expect integration tests to pass, if at least one unit test fails.

Organizing tests in layers like this offers us the possibility to force
ourselves to stay focus on the business value of the project, on the stuff that
matters. We do so by writing first the most abstract type of test, in our case,
an integration test. We draw the content of the test from the customer's pool
of stories, of use cases.

Starting with this commit, you will see different user stories (also called
use cases in UML terminology) which make the tic tac toe game be what it is.

Since we don't have any production code yet, we want to start with the simplest
story we can find in the pool of user stories. Note that in real life, these
user stories would be registered in a project management software for agile
project management, but, the current project being simplistic, we will just use
our imagination.

The first user story is "when the tictactoe game starts, the board of the game
is empty and the next player is X".

In this commit, we are writing the integration test for this story.

If you look at the code, you will notice how we use terminology from the quoted
user story above.

This is (simplistically seen), similar to Domain-Driven
Design (DDD), supposing that the user story was phrased by domain experts. One
important idea in DDD is to craft the code (in this case, the name of the
classes and the methods) according to the domain of the application, and this
is what we've done:

We have introduced two classes named `Board` and `Tictactoe` which holds the
board, and getter methods for the board and the next player. We then have
tested that the board is empty.

If you're new to TDD, writing the test code first may seem awkward, but once
you get used to it, you will notice that it helps in a few ways:

* first, it helps you keep focused on the task at hand and not invest time in
  implementing features you are not sure that you'll need
* the tests are forcing you to think about what you're doing
* most of the thinking is done while writing tests, afterwards you just try to
  make the tests pass, which most of the time means you read compiler errors
  and fix them one by one - you almost blindly follow what the compiler tells
  you to do
* when you have to fix a bug, the tests will tell you if you have fixed it
  without introducing any new regressions in the production code
* when you refactor some internal details of the classes, but the public API
  stays mostly the same, the tests will tell you if you have refactored it
  correctly, without introducing bugs
* when you can't recall how to use a component, a class, or how multiple
  classes or subsystems interact, you can use the tests as a documentation - it
  is, in fact, the most accurate documentation of the code which you have; it
  is so accurate, that it even lights green when you run the "documentation",
  meaning that this documentation truly documents the production code. You will
  learn to appreciate this source of documentation once you have to deal with
  millions of lines of code (not that LOCs are would be an accurate measure of
  project complexity)

Now, if you run the test target (`make test` in the console, in a Makefile
environment), you will notice that it gives errors, and this is normal, since
the production code does not yet exist.

You may think that you have accomplished nothing, but in fact, you have made an
important step towards the issues raised in the bullet points above, as well as
specified a slice of the public interface of the system.

You have, in fact, designed the software, in 4 lines of (testing) code.

The purpose of integration tests is to give you the rough direction of the path
you have to follow. Once you have made the integration test pass, you have
delivered business value.

However, the implementation details are going to be developed driven not by
integration tests, but by unit tests.

In the next commit, we are going to create some unit tests, which of course
will still fail, because we will still have no production code at all! You will
understand why this is like this in the next commit, and in the commit after
that, we will get it all rolling.