Continue the TDD cycle by writing unit tests
In the previous commit, we have written an use case as it could have been
specified by the stackholder in a specification language which we know: C++.
Indeed, we can look at C++ as a specification language, particularily when
talking about integration tests.

This time around, we are going to drill further into the abstractedness of the
tests (please recall how tests are organized into layers) by writing unit tests
for the methods which we have discovered would exist by looking at the
integration test.

Now it's the proper time to dissolve the file `test_unit_main.cpp` which we
have used as a conceptual placeholder for all unit tests. This file is going to
only contain the entry point (the main() function) provided by Boost.Test.

The first assertion in the integration test
`when_starting_the_game_the_board_is_empty_and_the_next_player_is_x` reveals
that there should be a method `Tictactoe::nextPlayer()`, for which we have
written the unit test `the_game_starts_with_x_active_player`.

If you pay attention, the assertion line in the integration test and in this
new unit test regarding the return value of `nextPlayer()` are identical, which
could make you think that we are writing duplicate code.

Fact is, even though the two lines are identical, the state of the SUT is
different: in the integration test, the `Tictactoe` instance has a board, in
the unit test it doesn't. Because we're probing the SUT in different states, we
are not, in fact, writing overspecified tests or brittle tests.

Brittle tests are a big no-go in TDD, as is overspecification,

The next test we have written in this commit is
`board_is_empty_after_creation`. This too looks almost like a duplicate, but
it's not, because we're testing the getter `board()` - we get to the `Board`
instance by using the getter, thus testing the integration of the two classes:
`Tictactoe` receives the `Board` as a dependency at construction time, and we
verify that the `Tictactoe` constructor keeps this object.

Now that we have unit tests which are directly related to some of the units
(classes, methods) we need, we are going to write some production code in order
to make these tests pass.

I encourage you to regenerate your workspace (command: `cmake .`) and attempt
to build the project (command: `make`). You will see the error:

    test_Board.cpp:6:19: fatal error: Board.h: No such file or directory

which is because in `test_Board.cpp` you have imagined you had a file
`Board.h`. So create it to get rid of this error.

This doesn't mean to also create the `Board` class, pretend for a moment you
wouldn't know what the code is supposed to look like and do the very minimum
required to get rid of the error: create an empty file `src/Board.h` - really,
just the empty file!

Now build the project again. This time it says:

    error: ‘Board’ was not declared in this scope

Now do what it almost begs for, open the empty file `src/Board.h` and declare
an empty class called `Board`, properly guarded against double-inclusion:

    #ifndef _BOARD_H_
    # define _BOARD_H_

    class Board {
    };

    #endif

Now run the build again. Again, the compiler tells you what to do:

    error: ‘class Board’ has no member named ‘isEmpty’

So go to the definition of that file and make it look like so:

    #ifndef _BOARD_H_
    # define _BOARD_H_

    class Board {
        public:
            bool isEmpty() const;
    };

    #endif

Finally, you should see how the unit test `test_Board.cpp` is compiled
successfully.

But the compiler starts to complain about the missing pieces for making
`test_Tictactoe.cpp` compile. Follow the required steps to make it compile,
just as you did for the class `Board`.

Now run the build process again. As we expected, there are no more compiler
errors left. However, there are linker errors. For instance, the linker would
complain about a missing symbol, like this:

    test_Board.cpp: undefined reference to `Board::isEmpty() const'

So open the file `src/CmakeLists.txt` and right after `main.cpp` add a new file
called `Board.cpp`. The line would look like:

    add_library(tictactoe_lib main.cpp Board.cpp)

Do not forget to regenerate the "workspace" every time you modify a
`CMakeLists.txt` file by running `cmake .`.

Then create an empty file `src/Board.cpp` and implement the methods the linker
complains about not finding.

Remember, in TDD, all you want to do is to make the tests pass, no more, no
less. Functionality does not matter, all it matters is to make the tests pass
as quickly as possible.

For instance, this implementation of Board.cpp makes the unit tests for `Board`
pass:

    #include <Board.h>

    bool Board::isEmpty() const {
        return true;
    }

Repeat the process for `Tictactoe.cpp` by repeatedly running the build system
(`make`) and fixing one linker issue at a time.

The toolset you're using, the build system, the compiler, the linker, they all
tell you in plain error and warning messages, what they expect from you, all
you have to do it to understand their messages and take appropiate action,
almost blindly.

If you've done it correctly, at some point you'll get a full build of both
executables, the executable running the unit tests and the one running the
integration test.

Now you can actually run them by running the `test` target (on the command
line, that is `make test`). You should see how all tests pass.

Finally, remove the `main.cpp` file from the tictactoe library in
`src/CMakeLists.txt`, as we don't longer need it: we now have code and we don't
need an empty placeholder like `main.cpp` any longer.

Now, you want to compare your work, with my work. First, let's clean the dirty
files:

    git clean -fdX

Now, stash your changes:

    git stash -u

Now go to the next lesson, but before starting to read the commit message of
the next lesson, run `git stash pop`. If there are any merge conflicts or just
differences, you will see how your implementation differs from mine.

It doesn't necessarily mean that your implementation was wrong, sometimes there
are multiple ways of solving a problem.

However, you should now get rid of your code or commit it to a sepparate
branch, so it won't conflict with how the project will evolve in the future:

    git reset --hard HEAD